test(contract): split api compatibility suite into groups
The single file ran ~57s, crossing vitest's 60s worker RPC deadline in CI. Split by concern (generics, member shape, readonly/accessors, nominal surfaces) over a shared fixture; the groups now run in parallel and no file exceeds ~20s.
